Invalidate book query after form submission
